Correct position of the menu when out of viewport
When a context menu is higher than the place above the coordinates you clicked it would go outside of the viewport at the top of the screen. This PR makes it so that it will appear central in on the screen in that case. If the menu is higher than the viewport it will stick to the top of the screen. Applied the same logic to the horizontal axis.